backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
anthropic==0.18.1
torch>=2.0.0
torchvision>=0.15.0
celery==5.3.6
redis==5.0.1
python-multipart==0.0.9
websockets==12.0
python-dotenv==1.0.0
